[test] remove context.py
With the context in place it is not sure if the installed or the local
package is tested. The local copy can always be tested from the top
folder because python adds the current directory to the path.
Alternatively, one can do a pip install with the --editable option.
Therefore, we prefer to remove the context at this point.