[ci] use python 3.10.4
The CI failed because of a mypy-error coming out of numpy with the error
message "Positional-only parameters are only supported in Python3.8 and
greater". Weirdly, this error came when using py3.10.7, but not with
py3.10.4. For now, we just change the patch version in the CI, but a
TODO comment + an issue was created to revisit this in the future.