rename _array -> _numpy_utils
It is unlikely that we move to another array type (+ algorithms).
Therefore, we change the name to a more descriptive one, stating
the actual purpose of this file (which basically wraps numpy for
our purposes). In case we ever move to another underlying implementation,
still, only the import has to be changed...